+++ src/global.d.ts
@@ -2,4 +2,3 @@ declare const MICROSOFT_CLARITY_ID: string;
 declare const DEFAULT_PAGE_SIZE: number;
 declare const __APP_VERSION__: string;
 declare const API_URL: string;
-declare const HYPERPARAMETERS: string[];