+++ vite.config.ts
@@ -10,7 +10,6 @@ const APP_VERSION = process.env['APP_VERSION'] || `${version}+dev`;
 const MICROSOFT_CLARITY_ID = process.env['MICROSOFT_CLARITY_ID'] || '';
 const API_URL =
     process.env['API_URL'] || 'http://substra-backend.node-1.com:8000';
-const HYPERPARAMETERS = process.env['HYPERPARAMETERS'] || [];
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -20,7 +19,6 @@ export default defineConfig({
             ? {
                   API_URL: `'${API_URL}'`,
                   MICROSOFT_CLARITY_ID: `'${MICROSOFT_CLARITY_ID}'`,
-                  HYPERPARAMETERS: HYPERPARAMETERS,
               }
             : {}),
         DEFAULT_PAGE_SIZE: '10',