+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Moved cypress to a dedicated package (#577)
 -   Use the name field of Compute Plans instead of tag (#578)
 -   Algo sorting by category is based on category names (#584)
+-   Removed concept of hyperparameters in favor of metadata (#580)
 
 ## [0.21.0] - 2022-05-09